requirements.txt
pyserial==3.5
pytchat==0.5.5